simple solution using dense matrixes
turns out that trying to be efficient using sparse matrix was actually
making the program slower. I assumed that map access could be made in
O(1), but it seems that it wasn't that simple. Not sure if the
bottleneck was the access time, or simply the Struct construct required
at each step in:

```go
if _, ok := grid[Coord{ii, jj}]; ok {
```

Anyway... Using a dense matrix works well for this example (1000x1000)
matrix, and runs much faster. Classical premature optimization.

Also spent a lot of time with the assumption that rule[0] was always
'.' (i.e. a region of 9 dark cells would remain dark every round).
Although this was the case on the sample input, the real input has
rule[0] == '#'. Solved that by defining a default_value for new
values (and assuming that the count is just valid for even iterations,
as odd ones will have infinite pixels on), but took me a while to
realize that issue.