after 5 months, finally found the solution for day24! :)
It took ages, but it was worth it. Happy that I didn't check for any
tips, and that this was done before the next advent!

The final solution was simple (possibly still a bit more complicated
than the optimal), but it took much longer than I expected to find it!